Reemplazo de los archivos originales por la version actualizada con el menu interactivo corregido
