+++ Clientes.ts
@@ -1,10 +1,14 @@
+import * as rls from 'readline-sync';
 import {Paciente} from "./Paciente";
+import {Consulta} from "./Consulta";
+import {GestorPrograma} from "./GestorPrograma";
+
 export class Cliente {
   private id: string;
   private nombre: string;
   private telefono: number;
   private dni : number;
-  private mascotas : Paciente[] = []
+  private listaMascotas : Paciente[] = []
   private numeroVisitas: number = 0;
   private esVIP: boolean = false;
   
@@ -55,11 +59,11 @@ export class Cliente {
   }
 
   public getMascotas():Paciente[]{
-    return this.mascotas;
+    return this.listaMascotas;
   }
 
-  public setMascotas(mascota:Paciente){
-    this.mascotas.push(mascota);
+  public setMascotas(listaMascotas:Paciente[]){
+    this.listaMascotas=(listaMascotas);
   }
 
   public getTelefono(): number {
@@ -84,5 +88,148 @@ export class Cliente {
 
   public setEsVIP(esVIP: boolean): void {
     this.esVIP = esVIP;
+  }
+
+  // METODOS MASCOTAS //
+
+  public agregarMascotaALaLista(mascota: Paciente): void {
+    this.listaMascotas.push(mascota);
+  }
+
+  public mostrarListaMascotas(){
+    console.log("Lista de Mascotas:");
+      this.listaMascotas.forEach((mascota, index) => {
+        console.log(`${index + 1}. Nombre: ${mascota.getNombre()}, Tipo: ${mascota.getEspecie()}`);
+      });
+  }
+
+  //CRUD MASCOTAS
+
+  public crearMascota(): void {
+    console.clear();
+        // Crear mascota
+        const nombre : string = rls.question("Ingrese el nombre de la Mascota: ");
+        const especie : string = rls.question("Ingrese la especie de su mascota: ");
+        const mascota = new Paciente(nombre, especie, this.id);
+
+        // Asignar mascota al cliente
+        this.agregarMascotaALaLista(mascota);
+        console.log(`Mascota ${mascota.getNombre()} agregado exitosamente al ${this.constructor.name}: ${this.nombre}`);
+        GestorPrograma.esperarEnter();
+}
+
+public listarMascotas(): void {
+  console.clear();
+  // Validar si tiene  mascotas
+  if (this.listaMascotas.length === 0) {
+      console.log(`No hay Mascotas registradas al cliente: ${this.nombre}.`);
+      if (rls.keyInYNStrict(`¿Desea crear una nueva Mascota?`)) {
+          // Ejecuta nuevamente el Metodo crearMascota si se selecciona 'Sí'
+          this.crearMascota();
+      } else {
+          return; // Salir porque no hay nada que listar
+      }
+  }
+
+  // Listar clientes
+  console.log(`Mascotas registradas al cliente: ${this.nombre}`);
+  this.mostrarListaMascotas();
+
+  // Esperar la interacción del usuario
+  GestorPrograma.esperarEnter();
+}
+
+public modificarMascota(): void {
+    console.clear();
+    let modificarOtro = true;
+    while (modificarOtro) {
+        this.listarMascotas();
+        let id: number = rls.questionInt("Ingrese el numero de Mascota que desee modificar: ") - 1;
+        if (id < 0 || id >= this.listaMascotas.length) {
+            console.log("Número inválido");
+            GestorPrograma.esperarEnter();
+        } else {
+            let mascota: Paciente = this.listaMascotas[id];
+            console.log(`Mascota: ${mascota.getNombre()} Especie: ${mascota.getEspecie()}`);
+            let nuevoNombre: string = rls.question("Ingrese nuevo nombre (o presione Enter para no modificar): ");
+            let nuevaEspecie: string = rls.question("Ingrese nuevo telefono (o presione 0 para no modificar): ");
+            if (nuevoNombre) {
+                mascota.setNombre(nuevoNombre);
+                console.log("Nombre Modificado Correctamente");
+            }
+            if (nuevaEspecie) {
+                mascota.setEspecie(nuevaEspecie);
+                console.log("Especie modificada correctamente");
+            }
+
+            modificarOtro = rls.keyInYNStrict(`¿Desea modificar otro cliente?`);
+        }
+    }
+    console.clear();
+}
+
+public eliminarMascota(): void {
+  console.clear();
+  let eliminarOtro = true;
+  while (eliminarOtro) {
+      this.listarMascotas();
+      let id: number = rls.questionInt("Ingrese el numero de la Mascota que desee eliminar: ") - 1;
+      if (id < 0 || id >= this.listaMascotas.length) {
+          console.log("Número inválido");
+          GestorPrograma.esperarEnter();
+      } else {
+          let mascota: Paciente = this.listaMascotas[id];
+          console.log(`Mascota: ${mascota.getNombre()} Especie: ${mascota.getEspecie()}`);
+          if (rls.keyInYNStrict(`¿Está seguro que desea eliminar la mascota ${mascota.getNombre()}?`)) {
+              this.listaMascotas = this.listaMascotas.filter(m => m !== mascota);
+              console.log(`${mascota.getNombre()} eliminado con éxito.`);
+              GestorPrograma.esperarEnter();
+          } else {
+              console.log("Eliminación cancelada.");
+              GestorPrograma.esperarEnter();
+          }
+      }
+
+      eliminarOtro = rls.keyInYNStrict(`¿Desea eliminar otra mascota?`);
+  }
+}
+
+public atenderMascota():void{
+  console.clear();
+  const mascota : Paciente = this.seleccionarMascota();
+  if (mascota!=null){
+    const consulta : Consulta = new Consulta(this,mascota);
+    const motivo : string = consulta.agregarMotivo();
+    const precio : number = consulta.calcularPrecio();
+    console.log(`Se registro la Consulta de su Mascota: ${mascota.getNombre()}\nMotivo: ${motivo}\nPrecio: ${precio}`);
+    this.incrementarVisitas();
+    GestorPrograma.esperarEnter();
+    console.clear();
+  }
+}
+
+    // Comprueba si el cliente tiene Mascotas
+    private seleccionarMascota(): Paciente {
+      console.clear();
+      const mascotas: Paciente[] = this.getMascotas();
+      if (mascotas.length === 0) {
+          console.log("No hay Mascotas creadas aún. Por favor, cree una para continuar.");
+          GestorPrograma.esperarEnter();
+          this.crearMascota();
+      }
+      this.listarMascotas();
+      console.clear();
+      const id = rls.questionInt("Ingrese el número de la Mascota que desea atender: ") - 1;
+
+      if (id < 0 || id >= mascotas.length) {
+          console.log("Número inválido. Regresando al menú principal.");
+          GestorPrograma.esperarEnter();
+      }
+
+      const mascota : Paciente = mascotas[id];
+      console.log(`Mascota seleccionada: ${mascota.getNombre()} (${mascota.getEspecie()})`);
+      GestorPrograma.esperarEnter();
+      console.clear();
+      return mascota;
   }
 }
