--- Veterinaria/README.md
@@ -1,15 +1,0 @@
-# Veterniaria
-
-## Descripcion
- El programa permite gestionar una red de veterinarias, con sus respectivos proveedores, clientes y pacientes.
-
-## Instrucciones de instalacion:
-
-1. Clonar el repositorio:
-   git clone https://github.com/Luciahernandosteckler/Veterinaria.git
-
-2. Navegar al directorio del proyecto.
-
-3. Instalar las dependencias necesarias:
-    npm install
-
