--- Veterinaria/main.ts
@@ -1,7 +1,0 @@
-import {GestorPrograma} from "./GestorPrograma";
-
-const gestor = new GestorPrograma("Gestor");
-
-gestor.opcionesGestorVeterinarias();
-
-
