--- Veterinaria/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "veterinaria",
-  "version": "1.0.0",
-  "main": "main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "readline-sync": "^1.4.10"
-  }
-}
